backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
google-generativeai
python-dotenv
python-multipart
numpy
pytest
httpx
